+++ DTCD-DevPanel-Console/src/components/ConsoleInputField.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="wrapper">
     <div class="console-prefix">
-      <i class="fas fa-chevron-right icon"/>
+      <i class="fas fa-chevron-right icon" />
     </div>
     <div
       ref="console"
@@ -44,11 +44,11 @@ export default {
     customConsoleCommands: ['log', 'info', 'warn', 'error'],
   }),
   watch: {
-    autocompletedSentence () {
+    autocompletedSentence() {
       this.appplyAutocompleteSentence();
     },
   },
-  mounted () {
+  mounted() {
     this.customConsoleCommands.forEach(command => {
       const defaultCommand = console[command].bind(console);
       console[command] = (...args) => {
@@ -58,30 +58,30 @@ export default {
     });
   },
   methods: {
-
-    addMessage (type, text) {
+    addMessage(type, text) {
       this.$emit('add-message', { type, text });
     },
 
-    clearMessages () {
+    clearMessages() {
       this.$emit('clear-messages');
     },
 
-    clearConsoleInput () {
+    clearConsoleInput() {
       this.expression = '';
       this.$refs.console.textContent = '';
     },
 
-    onInput () {
+    onInput() {
       this.updateExpression();
     },
 
-    onPaste (event) {
+    onPaste(event) {
       const text = event.clipboardData.getData('text/plain');
       document.execCommand('insertText', false, text);
     },
 
-    onEnterPress () {
+    onEnterPress() {
+      console.log(this.expression);
       if (this.expression === '') return;
 
       this.addExpressionToHistory();
@@ -96,7 +96,7 @@ export default {
       this.openAutocompleteTool(false);
     },
 
-    onArrowKeyPress (key) {
+    onArrowKeyPress(key) {
       const direction = key === 'up' ? 'prev' : 'next';
       if (this.autocompleteVisible) {
         this.selectAutocompleteSentence(direction);
@@ -105,7 +105,7 @@ export default {
       }
     },
 
-    onTabPress () {
+    onTabPress() {
       if (this.autocompleteVisible) {
         this.$emit('apply-sentence', true);
       } else {
@@ -113,36 +113,112 @@ export default {
       }
     },
 
-    updateExpression () {
-      this.expression = this.$refs.console.textContent.trim();
+    updateExpression() {
+      if (this.$refs.console.textContent.endsWith('(')) {
+        this.$refs.console.textContent = this.$refs.console.textContent + ')';
+        this.setCaretPosition(this.$refs.console.textContent.length - 1);
+      }
+      const caretPosition = this.getCaretIndex(this.$refs.console);
+      const leftParenthesisPosition = this.$refs.console.textContent.lastIndexOf(
+        '(',
+        caretPosition
+      );
+      const rightParenthesisPosition = this.$refs.console.textContent.indexOf(')', caretPosition);
+      if (leftParenthesisPosition != -1 || rightParenthesisPosition != -1) {
+        if (caretPosition > leftParenthesisPosition && caretPosition <= rightParenthesisPosition) {
+          const leftCommaPosition = this.$refs.console.textContent.lastIndexOf(
+            ',',
+            caretPosition - 1
+          );
+          const rightCommaPosition = this.$refs.console.textContent.indexOf(',', caretPosition);
+          if (
+            leftCommaPosition != -1 &&
+            leftCommaPosition > leftParenthesisPosition &&
+            leftCommaPosition < caretPosition &&
+            rightCommaPosition == -1
+          ) {
+            this.expression = this.$refs.console.textContent
+              .slice(leftCommaPosition + 1, rightParenthesisPosition)
+              .trim();
+          } else if (
+            rightCommaPosition != -1 &&
+            rightCommaPosition < rightParenthesisPosition &&
+            rightCommaPosition >= caretPosition &&
+            leftCommaPosition == -1
+          ) {
+            this.expression = this.$refs.console.textContent
+              .slice(leftParenthesisPosition + 1, rightCommaPosition)
+              .trim();
+          } else if (
+            rightCommaPosition != -1 &&
+            leftCommaPosition != -1 &&
+            rightCommaPosition < rightParenthesisPosition &&
+            leftCommaPosition > leftParenthesisPosition &&
+            leftCommaPosition < caretPosition &&
+            rightCommaPosition >= caretPosition
+          ) {
+            this.expression = this.$refs.console.textContent
+              .slice(leftCommaPosition + 1, rightCommaPosition)
+              .trim();
+          } else {
+            this.expression = this.$refs.console.textContent
+              .slice(leftParenthesisPosition + 1, rightParenthesisPosition)
+              .trim();
+          }
+        } else {
+          this.expression = this.$refs.console.textContent.trim();
+        }
+      } else {
+        this.expression = this.$refs.console.textContent.trim();
+      }
       this.$emit('expression-update', this.expression);
+      this.expression = this.$refs.console.textContent.trim();
     },
 
-    createExpressionFunction () {
+    setCaretPosition(pos) {
+      let sel = window.getSelection();
+      sel?.setPosition(this.$refs.console.childNodes[0], pos);
+    },
+    getCaretIndex(element) {
+      let position = 0;
+      const isSupported = typeof window.getSelection !== 'undefined';
+      if (isSupported) {
+        const selection = window.getSelection();
+        if (selection.rangeCount !== 0) {
+          const range = window.getSelection().getRangeAt(0);
+          const preCaretRange = range.cloneRange();
+          preCaretRange.selectNodeContents(element);
+          preCaretRange.setEnd(range.endContainer, range.endOffset);
+          position = preCaretRange.toString().length;
+        }
+      }
+      return position;
+    },
+
+    createExpressionFunction() {
       // TODO: Update after console logic revision
       return new Function(`
-        const robot = document._DataCAD;
-        robot.panelLiveDash = robot.systemGUID.getInstanceByName('PanelLiveDash');
+        const robot = Application.autocomplete;
         return (${this.expression});
       `);
     },
 
-    executeConsoleExpression () {
-      console.warn('Console disabled for revision');
-      // try {
-      //   const expression = this.createExpressionFunction();
-      //   this.addMessage('command', this.expression);
-      //   this.addMessage('log', expression());
-      // } catch (error) {
-      //   this.addMessage('error', error.message);
-      // }
+    executeConsoleExpression() {
+      // console.warn('Console disabled for revision');
+      try {
+        const expression = this.createExpressionFunction();
+        this.addMessage('command', this.expression);
+        this.addMessage('log', expression());
+      } catch (error) {
+        this.addMessage('error', error.message);
+      }
     },
 
-    addExpressionToHistory () {
+    addExpressionToHistory() {
       this.history.add(this.expression);
     },
 
-    restoreExpressionFromHistory (direction = 'prev') {
+    restoreExpressionFromHistory(direction = 'prev') {
       const expression = this.history[direction]();
       if (expression) {
         this.$refs.console.textContent = expression;
@@ -151,22 +227,22 @@ export default {
       this.updateExpression();
     },
 
-    openAutocompleteTool (isOpen) {
+    openAutocompleteTool(isOpen) {
       this.$emit(isOpen ? 'autocomplete-open' : 'autocomplete-close');
     },
 
-    appplyAutocompleteSentence () {
+    appplyAutocompleteSentence() {
       this.$refs.console.textContent = this.autocompletedSentence;
       this.updateExpression();
       this.$refs.console.focus();
       this.moveCaretToEndPosition();
     },
 
-    selectAutocompleteSentence (direction) {
+    selectAutocompleteSentence(direction) {
       this.$emit('select-sentence', direction);
     },
 
-    moveCaretToEndPosition () {
+    moveCaretToEndPosition() {
       const range = document.createRange();
       range.selectNodeContents(this.$refs.console);
       range.collapse(false);
@@ -174,7 +250,6 @@ export default {
       selection.removeAllRanges();
       selection.addRange(range);
     },
-
   },
 };
 </script>
@@ -184,8 +259,8 @@ export default {
 
 .wrapper {
   display: flex;
-  background-color: #F5F5F5;
-  box-shadow: 0 0 3px rgba(0, 0, 0, .5);
+  background-color: #f5f5f5;
+  box-shadow: 0 0 3px rgba(0, 0, 0, 0.5);
   position: relative;
 
   .console-prefix {