+++ CHANGELOG.md
@@ -1,6 +1,18 @@
-# Changelog
+# **CHANGELOG**
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Changed
+
+- name of plugin from DevPanel-Console to DevPanel_Console
+- the way autcomplete works
+- makefile
+
+### Removed
+
+- autocomplete hardcode file